Use the extend method of the transport manager instead of actually extending it
